Add basic styling
