+++ src/index.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
 
-
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
 );
