+++ src/interface/PreviewBox.module.css
@@ -8,5 +8,5 @@
   background-color: aliceblue;
   padding: 0.5em;
   display: inline-block;
-
+  
 }
