+++ src/interface/Toolbar.module.css
@@ -1,6 +1,6 @@
 .toolbar {
-  display: flex;
-  border-bottom: solid 1.5px rbga(128, 128, 128, 0.5); 
-  padding: 10px 0;
-  margin: 0 0 10px 0;
+  position: relative;
+  padding: 1px 18px 17px;
+  margin: 0px -20px 20px;
+  border-bottom: 2px solid rgb(238, 238, 238);
 }
