+++ src/pages/Editor.module.css
@@ -1,0 +1,7 @@
+.notes {
+  background-color: ghostwhite;
+  max-width: 42em;
+  margin: 20px auto;
+  padding: 20px;
+  border: none;
+}
